address cleanup comments
