+++ examples/tutorials/config_sample_2pop.yml
@@ -12,7 +12,6 @@ initial_conditions:
   method: SetInitialConditions
   initial_conditions_file: model_input/ic_2pop.csv
   allow_missing_compartments: TRUE
-  allow_missing_compartments: TRUE
 
 compartments:
   infection_stage: ["S", "E", "I", "R"]
